domain/keys: rename to `key_manipulation`
