+++ dexios-domain/src/lib.rs
@@ -4,7 +4,7 @@ pub mod erase;
 pub mod erase_dir;
 pub mod hash;
 pub mod hasher;
-pub mod keys;
+pub mod key_manipulation;
 pub mod overwrite;
 pub mod pack;
 pub mod storage;
